requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
databases[asyncpg]
python-dotenv
pydantic-settings
psycopg2-binary
pytest-env
rich
asgi-correlation-id
python-json-logger
httpx
python-jose
python-multipart
bcrypt
pytest-mock
alembic
aiofiles
b2sdk
pyfakefs
